requirements.txt
black
isort
pre-commit
pylint
python-gitlab==2.6.0
